PARQUET-2948: Modify the test case.
